+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.5.0
+
+### Fixed
+
+- `lisp-comment-dwim-region` better handles mixed content, it can
+  simultaneously (un)comment a mix of text & s-expressions
+
 ## 1.4.0
 
 ### Added